email_app_simulation/backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
certifi
pymongo
